refactor: rename typography utils to rich-text
