+++ packages/plugins/text/src/utils/rich-text.ts
